Refactor based on feedback
