+++ test/testing_images/def50.bmp
